crbuild: (hidden) setting to put user data dir in disk image.
No exposed option, and things are hard-coded, but will create
(and format) a disk image, and then mount it, using loopback
device, into a folder.

Will replace ${user_data_img_dir} in command-line args with the
path to the mounted user data folder.

Need to come up with a way to simply make this a command-line
arg to crbuild and require no modifications to collections.json.