crbuild: Using cgdb/gdb and test launcher jobs
The compile/test jobs are no longer the same value. in collections.json
you can use ${testjobs} to use a separate value, currently hard-coded
to 8. If set to None then that option will be removed before running
the test to assume the default behavior of the test.

Can also use wither cgdb or gdb --tui. Currently hard-coded to cgdb.